requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==21.2.0
black==21.11b1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
configparser==5.2.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.1.6
django-ckeditor==6.1.0
django-filter==21.1
django-hashid-field==3.3.3
django-jalali==4.3.0
django-jalali-date==1.0.0
django-js-asset==1.2.2
django-minio-storage==0.3.10
django-mptt==0.13.4
django-reservation==0.2.9
django-widget-tweaks==1.4.9
django-zarinpal==1.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
gunicorn==20.1.0
hashids==1.3.1
idna==3.3
isodate==0.6.1
isort==5.10.1
jdatetime==3.6.4
kavenegar==1.1.2
lxml==4.7.1
mccabe==0.6.1
minio==6.0.2
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.1
redis==4.0.2
regex==2021.11.10
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
soap==0.0.1rc0
sqlparse==0.4.1
suds==1.0.0
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
zeep==3.4.0
